fix: memperbaiki error akibat null ketika melakukan request api

+++ catalog/src/main/java/apap/tk/catalog/dto/request/UpdateCatalogDTO.java
@@ -1,6 +1,8 @@
 package apap.tk.catalog.dto.request;
 
 
+import java.math.BigInteger;
+
 import apap.tk.catalog.model.Category;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -11,7 +13,7 @@ import lombok.NoArgsConstructor;
 @Data
 public class UpdateCatalogDTO {
     private String productName;
-    private int price;
+    private BigInteger price;
     private String productDescription;
     private int stock;
     private byte[] image;
